IRQ update: the code now reaches the handlers(!) but still fails to handle it. More work to do.

+++ root/orbit-kernel/irq.h
@@ -1,6 +1,6 @@
 #ifndef IRQ_H
 #define IRQ_H
 
-double PIT_IRQ_timer_get_current_time_since_boot(void);
+int PIT_IRQ_timer_get_current_time_since_boot(void);
 
 #endif
